requirements.txt
PyQt5==5.15.11
matplotlib==3.10.3
mysql-connector-python==8.0.33
bcrypt==4.3.0
python-dotenv==1.1.1